conexao com o banco de dados, remocao de campos do front, modificacao no Program.cs

+++ CadastroAcoes/Model/Repository/IUserClientRepository.cs
@@ -9,6 +9,6 @@ namespace Model.Repository
     public interface IUserClientRepository
     {
         Task CreateAsync(UserClient user);
-        Task<UserClient?> GetByCfpAsync(int cfp);
+        Task<UserClient?> GetByCfpAsync(long cfp);
     }
 }
